Fix: changed tabs to pages
