+++ frontend/pages/10_settings_UI.py
@@ -1,0 +1,5 @@
+import streamlit as st
+from components.process_pdf import display_backend_status
+
+
+display_backend_status()
