+++ frontend/pages/3_tables_UI.py
@@ -1,0 +1,7 @@
+import streamlit as st
+import asyncio
+import httpx
+import os
+
+st.write("📋 Extract Tables")
+
