+++ frontend/pages/6_translate_UI.py
@@ -1,0 +1,4 @@
+import streamlit as st
+import logging
+
+st.write("📄 Translate Content")
